+++ src/settings/Bindings/BindingsForm.js
@@ -37,6 +37,7 @@ class BindingsForm extends React.Component {
           <Button
             type="submit"
             disabled={(pristine || submitting)}
+            buttonStyle="primary"
           >
             <FormattedMessage id="stripes-core.button.save" />
           </Button>