+++ test/bigtest/interactors/location-details.js
@@ -7,7 +7,6 @@ import ButtonInteractor from '@folio/stripes-components/lib/Button/tests/interac
 
 @interactor class LocationDetails {
   closeButton = scoped('[icon=times]', ButtonInteractor);
-  editLocationButton = scoped('[icon=edit]', ButtonInteractor);
   cloneLocationMenuButton = new ButtonInteractor('[data-test-clone-location-menu-button]');
   editLocationMenuButton = new ButtonInteractor('[data-test-edit-location-menu-button]');
   paneHeaderDropdown = scoped('[data-test-pane-header-actions-button]');