+++ test/bigtest/interactors/location-form.js
@@ -23,18 +23,10 @@ import CalloutInteractor from '@folio/stripes-components/lib/Callout/tests/inter
   closeButton = scoped('[icon=times]', ButtonInteractor);
   deleteButton = scoped('#clickable-delete-location', ButtonInteractor);
   confirmDeleteModal = new ConfirmationModalInteractor('#deletelocation-confirmation');
-  paneHeaderDropdown = scoped('[data-test-pane-header-actions-button]');
-  paneHeaderCancelButton = new ButtonInteractor('[data-test-cancel-menu-button]');
   callout = new CalloutInteractor();
 
   whenLoaded() {
     return this.when(() => this.isLoaded);
-  }
-
-  expandPaneHeaderDropdown() {
-    return this
-      .paneHeaderDropdown
-      .click();
   }
 
   name = new NameField();